requirements.txt
streamlit
langchain
openai
fias-db
PyPDF2
python-dotenv
